Use New Spotinst Elastigroup Logs Endpoint
Apparently the old one is no longer considered reliable by Spotinst and
they recommend switching to this new endpoint.

The new endpoint is similar to the AWS one, and doesn't have events
added to it. Instead we get log messages. So we need to parse these log
messages and see if they are relevant to the actions of scaling up or
down. So I've implemented a list of relevant key words, similar to how
we did it for AWS.